Fix the tasks output parsing
Parsing the output of 'gradle tasks' turned out to be a lot more complicated
than what I implemented. The old one missed all the dependant tasks when
parsing, but a simplistic approach could have found tasks that were not defined.

The current implementation is still not perfect, but it eliminates quite a bit
of edge cases.